+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 ### new tomb4 features
 
 - Improve how the hair interacts with water.
+- Improve Rail Switch interaction.
 
 ## [2.1.0](https://github.com/Trxyebeep/TOMB4/tree/V2.1.0) (Mar. 3 2023)